build: :construction: Creacion de las clases AdministradorInventario y ControlInventario
Creación de la clase AdministradorInventario para usarse en la vista del Administrador de inventario así como la clase ControlInventario donde se encuentran todas las consultas a la base de datos que son utilizadas por la clase AdminitradorInventario

+++ ventas/src/main/java/com/vfarma/Modelo/Producto.java
@@ -43,7 +43,7 @@ public class Producto {
         this.nombreProducto = nombreProducto;
     }
 
-    public void colocarPrecioProducto(int precioProducto) {
+    public void colocarPrecioProducto(float precioProducto) {
         this.precioProducto = precioProducto;
     }
